added basic action listener
