+++ src/helper/messageFormat.ts
@@ -1,8 +1,6 @@
-import { Middleware, SlackEventMiddlewareArgs } from '@slack/bolt';
-import { Users, Conversations, EventContext } from '../model';
-import * as Helpers from '../helper';
+import { SayArguments, Context } from '@slack/bolt';
+import { EventContext } from '../model';
 import _ from 'lodash';
-import moment from 'moment-timezone';
 
 const TIME_DISPLAY_FORMAT = 'MMM Do ddd, h:mm a';
 
@@ -12,20 +10,20 @@ const dateToUl = (date: EventContext.LocalDateReference) => {
     }`;
 };
 
-export const userConfirmationMsgBox = (context: EventContext.MessageTimeContext) => {
-    const dateRef = context.content[0];
+export const userConfirmationMsgBox = (timeContext: EventContext.MessageTimeContext, channelContext: Context) => {
+    const dateRef = timeContext.content[0];
 
     let dateDisplayString = '';
 
-    if (context.content.length > 0) {
-        for (let i = 0; i < context.content.length; i++) {
-            dateDisplayString = dateDisplayString.concat(dateToUl(context.content[i]));
+    if (timeContext.content.length > 0) {
+        for (let i = 0; i < timeContext.content.length; i++) {
+            dateDisplayString = dateDisplayString.concat(dateToUl(timeContext.content[i]));
         }
     } else {
         dateDisplayString = dateToUl(dateRef);
     }
 
-    const messageBlock = {
+    const messageBlock: SayArguments = {
         text: 'convert date confirmation',
         blocks: [
             {
@@ -54,7 +52,7 @@ export const userConfirmationMsgBox = (context: EventContext.MessageTimeContext)
                             text: 'Yes Please',
                         },
                         style: 'primary',
-                        value: 'btn_approve',
+                        value: JSON.stringify(channelContext),
                         action_id: 'convert_date',
                     },
                     {
