+++ src/middleware/index.ts
@@ -1,1 +1,2 @@
 export * from './messageListener';
+export * from './actionListener';
